+++ packages/icestark/src/AppRoute.tsx
@@ -86,8 +86,12 @@ export default class AppRoute extends React.Component<AppRouteProps, AppRouteSta
     };
     const { loadScriptMode, runtime } = props;
 
-    if (loadScriptMode !== 'fetch' && runtime) {
-      console.error('[icestark] runtime option can only be used when loadScriptMode is set to "fetch"');
+    if (runtime) {
+      if (loadScriptMode && loadScriptMode !== 'fetch') {
+        console.error('[icestark] runtime option can only be used when loadScriptMode is set to "fetch"');
+      } else if (!loadScriptMode) {
+        console.log('[icestark] Runtime option detected but loadScriptMode not set - automatically switching to "fetch" mode');
+      }
     }
   }